strucutre refactor for independent models and providers
